+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/gitlab
 
+## 0.15.24
+
+### Patch Changes
+
+- Updated dependencies [1f14793]
+- Updated dependencies [2619d79]
+  - @cat-factory/contracts@0.230.1
+  - @cat-factory/kernel@0.232.0
+
+## 0.15.23
+
+### Patch Changes
+
+- Updated dependencies [e7e4404]
+  - @cat-factory/contracts@0.230.0
+  - @cat-factory/kernel@0.231.0
+
 ## 0.15.22
 
 ### Patch Changes