+++ backend/packages/prompt-fragments/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/prompt-fragments
 
+## 0.15.56
+
+### Patch Changes
+
+- Updated dependencies [1f14793]
+  - @cat-factory/contracts@0.230.1
+
+## 0.15.55
+
+### Patch Changes
+
+- Updated dependencies [e7e4404]
+  - @cat-factory/contracts@0.230.0
+
 ## 0.15.54
 
 ### Patch Changes