+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/provider-s3
 
+## 0.2.290
+
+### Patch Changes
+
+- Updated dependencies [1f14793]
+- Updated dependencies [2619d79]
+  - @cat-factory/kernel@0.232.0
+
+## 0.2.289
+
+### Patch Changes
+
+- Updated dependencies [e7e4404]
+  - @cat-factory/kernel@0.231.0
+
 ## 0.2.288
 
 ### Patch Changes