+++ backend/packages/workspaces/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/workspaces
 
+## 0.21.43
+
+### Patch Changes
+
+- Updated dependencies [1f14793]
+- Updated dependencies [2619d79]
+  - @cat-factory/contracts@0.230.1
+  - @cat-factory/kernel@0.232.0
+
+## 0.21.42
+
+### Patch Changes
+
+- Updated dependencies [e7e4404]
+  - @cat-factory/contracts@0.230.0
+  - @cat-factory/kernel@0.231.0
+
 ## 0.21.41
 
 ### Patch Changes